refactor(remote): split navigator creation and remote registration hooks
