+++ packages/lib/src/spatial-navigation/components/Root.tsx
@@ -3,6 +3,7 @@ import { ParentIdContext } from '../context/ParentIdContext';
 import { SpatialNavigatorContext } from '../context/SpatialNavigatorContext';
 import { useBeforeMountEffect } from '../hooks/useBeforeMountEffect';
 import { useCreateSpatialNavigator } from '../hooks/useCreateSpatialNavigator';
+import { useRemoteControl } from '../hooks/useRemoteControl';
 import { OnDirectionHandledWithoutMovement } from '../SpatialNavigator';
 
 const ROOT_ID = 'root';
@@ -40,8 +41,10 @@ export const SpatialNavigationRoot = ({
 
   const spatialNavigator = useCreateSpatialNavigator({
     onDirectionHandledWithoutMovementRef,
-    isActive,
   });
+
+  useRemoteControl({ spatialNavigator, isActive });
+
   useBeforeMountEffect(() => {
     spatialNavigator.registerNode(ROOT_ID, { orientation: 'vertical' });
     return () => spatialNavigator.unregisterNode(ROOT_ID);
